requirements.txt
discord.py
google-generativeai
Pillow
PyNaCl
gTTS
SpeechRecognition
pydub
sqlite3‎
